Fixed styling with submit modal and reveal phase.
