+++ src/features/vote/SubmitModal.tsx
@@ -9,7 +9,7 @@ import {
 
 import Modal from "common/components/modal";
 import Button from "common/components/button";
-import { ModalWrapper } from "./styled/ActiveRequestsForm.styled";
+import { ModalWrapper } from "./styled/SubmitModal.styled";
 import { SubmitModalState, Summary, FormData } from "./CommitPhase";
 import { UnlockedIcon, LockedIconCommitted } from "assets/icons";
 import {
